# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "업무 네비게이터 API",
    "description": "학교 업무 담당자의 질문에 다음 할 일과 근거 문서를 돌려주는 API. 지금은 모든 엔드포인트가 예시 데이터로 동작한다.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "업무 흐름 조회",
        "operationId": "get_workflow_api_v1_workflows__workflow_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "examples": [
                "science_competition"
              ],
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/steps/{step_id}/complete": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "단계 완료 처리",
        "description": "완료 처리한 뒤 갱신된 업무 흐름 전체를 돌려준다. 프론트는 응답으로 화면을 그대로 다시 그리면 되고, 따로 조회하지 않아도 된다.",
        "operationId": "complete_step_api_v1_workflows__workflow_id__steps__step_id__complete_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Step Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/feedback": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "실제 업무가 등록된 흐름과 다를 때 알린다",
        "operationId": "post_feedback_api_v1_workflows__workflow_id__feedback_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "프론트가 분기에 쓰는 코드. 사람이 읽는 문구는 message를 쓴다.",
            "examples": [
              "workflow_not_found"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "요청한 업무를 찾을 수 없습니다."
            ]
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "검증 실패(422)일 때 어떤 필드가 왜 틀렸는지.",
            "examples": [
              [
                {
                  "field": "body.query",
                  "reason": "String should have at least 1 character"
                }
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "4xx/5xx 응답의 공통 형태. 프론트는 error.code로 분기한다."
      },
      "FeedbackRequest": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/FeedbackType"
          },
          "after_step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Step Id",
            "description": "기준이 되는 단계. 이 단계 다음에 차이가 있다는 뜻이다. null이면 워크플로 맨 앞을 가리킨다.",
            "examples": [
              "2"
            ]
          },
          "suggested_step_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Step Name",
            "description": "missing_step일 때 실제로 했던 일의 이름.",
            "examples": [
              "개인정보 동의"
            ]
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description",
            "description": "담당자가 직접 쓴 설명. 이후 워크플로 개선 검토에 쓴다."
          },
          "query_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Id",
            "description": "이 피드백을 유발한 /query 응답이 있으면 그 식별자."
          }
        },
        "type": "object",
        "required": [
          "type",
          "description"
        ],
        "title": "FeedbackRequest"
      },
      "FeedbackResponse": {
        "properties": {
          "feedback_id": {
            "type": "string",
            "title": "Feedback Id",
            "examples": [
              "fb_01HZX4A2"
            ]
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "type": {
            "$ref": "#/components/schemas/FeedbackType"
          },
          "diff": {
            "$ref": "#/components/schemas/WorkflowDiff"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "담당자에게 보여줄 접수 확인 문구.",
            "examples": [
              "의견이 접수되었습니다. 워크플로 개선 검토에 반영됩니다."
            ]
          }
        },
        "type": "object",
        "required": [
          "feedback_id",
          "workflow_id",
          "type",
          "diff",
          "message"
        ],
        "title": "FeedbackResponse"
      },
      "FeedbackType": {
        "type": "string",
        "enum": [
          "missing_step",
          "unnecessary_step",
          "wrong_order",
          "wrong_document",
          "other"
        ],
        "title": "FeedbackType",
        "description": "실제 업무가 등록된 워크플로와 다를 때, 어떻게 다른지."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StepCompleteRequest": {
        "properties": {
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "default": true
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "담당자가 남기는 메모. 업무 Trace에 함께 기록된다."
          }
        },
        "type": "object",
        "title": "StepCompleteRequest",
        "description": "단계 완료 처리.\n\ncompleted를 false로 보내면 완료를 되돌린다. 잘못 누른 경우를 위해\n되돌리기도 같은 엔드포인트로 처리한다."
      },
      "StepStatus": {
        "type": "string",
        "enum": [
          "completed",
          "current",
          "pending"
        ],
        "title": "StepStatus",
        "description": "워크플로 한 단계의 진행 상태."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowDetail": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "workflow_id",
          "name",
          "steps"
        ],
        "title": "WorkflowDetail",
        "description": "업무 흐름 시각화에 필요한 전부.\n\nsteps는 진행 순서대로 정렬해서 보낸다. 프론트가 다시 정렬하지 않아도 된다."
      },
      "WorkflowDiff": {
        "properties": {
          "expected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expected",
            "examples": [
              [
                "학생 선발",
                "참가 신청"
              ]
            ]
          },
          "reported": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reported",
            "examples": [
              [
                "학생 선발",
                "개인정보 동의",
                "참가 신청"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "expected",
          "reported"
        ],
        "title": "WorkflowDiff",
        "description": "등록된 흐름과 담당자가 말한 실제 흐름의 차이.\n\n프론트는 두 배열을 나란히 놓아 예상 / 실제로 보여줄 수 있다."
      },
...
      "WorkflowStep": {
        "properties": {
          "step_id": {
            "type": "string",
            "title": "Step Id",
            "examples": [
              "3"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "참가 신청"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/StepStatus"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "학생 명단을 첨부해 참가 신청서를 제출한다."
            ]
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "완료 처리된 시각. status가 completed일 때만 값이 있다."
          },
          "document_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Document Ids",
            "description": "이 단계에서 참고하거나 작성하는 문서."
          }
        },
        "type": "object",
        "required": [
          "step_id",
          "name",
          "status"
        ],
        "title": "WorkflowStep",
        "description": "워크플로의 한 단계."
      },
...
    }
  }
}